third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-template-get-set-expected.txt

This test checks that grid-template is properly parsed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-template set through CSS.
PASS window.getComputedStyle(gridWithDefaultTemplate).getPropertyValue('grid-template') is "none"
PASS window.getComputedStyle(gridWithSingleStringTemplate).getPropertyValue('grid-template') is "\"area\""
PASS window.getComputedStyle(gridWithTwoColumnsTemplate).getPropertyValue('grid-template') is "\"first second\""
PASS window.getComputedStyle(gridWithTwoRowsTemplate).getPropertyValue('grid-template') is "\"first\" \"second\""
PASS window.getComputedStyle(gridWithSpanningColumnsTemplate).getPropertyValue('grid-template') is "\"span span\""
PASS window.getComputedStyle(gridWithSpanningRowsDotTemplate).getPropertyValue('grid-template') is "\"span\" \".\""
PASS window.getComputedStyle(gridWithDotColumn).getPropertyValue('grid-template') is "\"header .\" \"footer .\""
Test grid-template: initial
PASS window.getComputedStyle(element).getPropertyValue('grid-template') is "\"foobar\""
Test grid-template: inherit
PASS window.getComputedStyle(parentElement).getPropertyValue('grid-template') is "\"foo bar\""
PASS window.getComputedStyle(element).getPropertyValue('grid-template') is "\"foo bar\""
Test invalid grid-template values.
PASS window.getComputedStyle(element).getPropertyValue('grid-template') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template') is "none"

FIXME: We currently don't validate that the named grid areas are <indent>.
FAIL window.getComputedStyle(element).getPropertyValue('grid-template') should be none. Was "nav-up".
PASS successfullyParsed is true

TEST COMPLETE
